Add the stuffz
This now builds a docker image with BIND9 and a config to allow clients
to resolve names.

It will download a copy of named.root so we know we have a fresh copy of
the hints file.

gen-config is a small script that generates BIND config from a template
using Python string formatting so it doesn't require any third party
dependencies. It is used to put the GOODCLIENTS environment option into
the BIND config file.